LibJS: Allow usage of locals if function parameters have default values
Initially, the usage of local variables for parameters had to be
disabled when default values were used because there was a bug that
didn't allow to correctly find the scope to which identifiers used
within the default parameter expression belonged, and hence correctly
identify if a variable can be local. However, this bug was fixed in
2f85faef0ff1cc823d4021dd3f3af589d6fc850f, so now this restriction
is no longer needed.